MCP-Orchestrator/backend/requirements.txt
# Django and REST Framework
Django>=4.0,<5.0
djangorestframework>=3.14.0,<4.0
django-filter>=23.0,<24.0
django-cors-headers>=4.0.0,<5.0

# Environment variables
python-dotenv>=1.0.0,<2.0

# Channels for WebSocket
channels>=4.0.0,<5.0
daphne>=4.0.0,<5.0
channels-redis>=4.1.0,<5.0  # Optional: for production use with Redis

# SSH and Crypto
paramiko>=3.0.0,<4.0
cryptography>=42.0.0,<43.0

# ASGI server
uvicorn>=0.23.0,<1.0
gunicorn>=21.0.0,<22.0

# Database
# psycopg2-binary>=2.9.5,<3.0  # Uncomment for PostgreSQL

# Development tools
black>=23.0,<24.0
flake8>=6.0.0,<7.0
pytest>=7.0.0,<8.0
pytest-django>=4.5.0,<5.0
